Split common parts of terminal service out
This is needed to fix the layer issue. This commit is a WIP.
